% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add.date}
\alias{add.date}
\title{add calibrated distributions to a plot.}
\usage{
add.date(
  age,
  error,
  depth,
  cc = 1,
  postbomb = FALSE,
  reservoir = c(),
  calibt = c(),
  prob = 0.95,
  threshold = 1e-04,
  BCAD = FALSE,
  ex = 1,
  draw.hpd = TRUE,
  hpd.lwd = 2,
  hpd.col = rgb(0, 0, 1, 0.7),
  mirror = FALSE,
  up = TRUE,
  on.axis = 1,
  col = rgb(0, 0, 1, 0.5),
  border = rgb(0, 0, 1, 0.5)
)
}
\arguments{
\item{age}{Mean of the uncalibrated C-14 age.}

\item{error}{Error of the uncalibrated C-14 age.}

\item{depth}{Depth of the date}

\item{cc}{Calibration curve for C-14 dates (1, 2, 3, or 4, or, e.g., "IntCal20", "Marine20", "SHCal20", "nh1", "sh3", or "mixed").}

\item{postbomb}{Whether or not this is a postbomb age. Defaults to FALSE.}

\item{reservoir}{Reservoir age, or reservoir age and age offset.}

\item{calibt}{Calibration based on the student-t distribution. By default, the Gaussian distribution is used (\code{calibt=FALSE}). To use the student-t distribution, provide two parameters such as \code{calibt=c(3,4)}.}

\item{prob}{Probability confidence intervals (between 0 and 1).}

\item{threshold}{Report only values above a threshold. Defaults to \code{threshold=1e-6}.}

\item{BCAD}{Use BC/AD or cal BP scale (default cal BP).}

\item{ex}{Exaggeration of the height of the distribution}

\item{draw.hpd}{Whether or not to draw the hpd ranges as a line}

\item{hpd.lwd}{Width of the line of the hpd ranges}

\item{hpd.col}{Colour of the hpd rectangle}

\item{mirror}{Plot distributions mirrored, a bit like a swan. Confuses some people but looks nice to the author so is the default.}

\item{up}{If mirror is not supplied, the distribution can be plotted up or down, depending on the direction of the axis.}

\item{on.axis}{Which axis to plot on. Defaults to 'x' or 1, but can be set to 'y' or 2.}

\item{col}{Colour of the inside of the distribution}

\item{border}{Colour of the border of the distribution}
}
\description{
Add individual calibrated dates to a plot.
}
\examples{
  plot(0, xlim=c(500,0), ylim=c(0, 2))
  add.date(130, 20, depth=1) 
}
