% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{contamination}
\alias{contamination}
\title{The impact of contamination on a radiocarbon age}
\usage{
contamination(y, sdev = c(), fraction, F14C, F14C.er = 0, decimals = 5)
}
\arguments{
\item{y}{the true radiocarbon age}

\item{fraction}{Relative amount of contamination. Must be between 0 and 1}

\item{F14C}{the F14C of the contamination. Set at 1 for carbon of modern radiocarbon age, at 0 for 14C-free carbon, or anywhere inbetween.}

\item{F14C.er}{error of the contamination. Defaults to 0.}

\item{er}{the error of the true radiocarbon age}
}
\value{
The observed radiocarbon age and error
}
\description{
Given a certain radiocarbon age, calculate the observed impact of contamination with an amount of material of a different 14C ratio
}
\examples{
contamination(5000, 20, .01, 1)
plot impacts of different amounts of contamination with modern carbon:
real.14C <- seq(0, 50e3, length=200)
contam <- seq(0, .2, length=100) # 0 to 20\% contamination
contam.col <- rainbow(length(contam))
plot(0, type="n", xlim=range(real.14C), xlab="real 14C age", ylim=range(real.14C), ylab="observed 14C age")
legend("topleft", legend=seq(0, .2, length=6), text.col=rainbow(6), bty="n")
for(i in 1:length(contam))
  lines(real.14C, contamination(real.14C, c(), contam[i], 1, decimals=5), col=contam.col[i])
}
\author{
Maarten Blaauw
}
