% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntCal.R
\name{copyCalibrationCurve}
\alias{copyCalibrationCurve}
\title{Copy a calibration curve}
\usage{
copyCalibrationCurve(cc = 1, postbomb = FALSE)
}
\arguments{
\item{cc}{Calibration curve for 14C dates: \code{cc=1} for IntCal20 (northern hemisphere terrestrial), \code{cc=2} for Marine20 (marine),
\code{cc=3} for SHCal20 (southern hemisphere terrestrial). Alternatively, one can also write, e.g., "IntCal20", "Marine13".}

\item{postbomb}{Use \code{postbomb=TRUE} to get a postbomb calibration curve (default \code{postbomb=FALSE}).}
}
\value{
The calibration curve (invisible).
}
\description{
Copy one of the calibration curves into memory. Renamed to ccurve, and copyCalibrationCurve will become obsolete
}
\details{
Copy the radiocarbon calibration curve defined by cc into memory.
}
