% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcal_json.R
\name{intcal.extract.record}
\alias{intcal.extract.record}
\title{Extract an IntCal20 record}
\usage{
intcal.extract.record(i)
}
\arguments{
\item{i}{The IntCal record, in the order as they appear in the IntCal20 json file (140 entries). Must be a single integer between 1 and 140.}
}
\value{
The most relevant information for each record (where available, the record and site names, intcal series and division number, country, longitude, latitude, taxon, dois of publications, notes, calendar information (e.g., dendrochronology) and radiocarbon data)
}
\description{
Extract one of the 140 records contributing to the IntCal20 calibration curves
}
\examples{
 record_1 <- intcal.extract.record(1)
}
