% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcal_json.R
\docType{data}
\name{intcal}
\alias{intcal}
\title{IntCal20 json file}
\format{
## `intcal`
A list with six main entries:
\describe{
  \item{json_application}{IntChron project name}
  \item{records}{a list with 139 entries for each IntCal dataset}
  \item{project_series_list}{a list with 5 entries: IntCal20, Marine20, SHCal20, a list of the underlying datasets, and a GICC vs IntCal20 comparison}
  \item{parameters}{an empty list}
  \item{bibliography}{a list with 141 bibliography entries}
  \item{options}{a list of 17 options (not used)}
}
}
\source{
<https://intchron.org/archive/IntCal/IntCal20/index.json>
}
\usage{
intcal
}
\description{
The IntCal20 calibration curves and their underpinning data. 
This is based on a json file produced by Prof. Christopher Bronk Ramsey, University of Oxford.
}
\keyword{datasets}
