% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcal_json.R
\name{intcal.read.data}
\alias{intcal.read.data}
\title{Read data underlying the IntCal curves.}
\usage{
intcal.read.data(
  fname = url("https://intchron.org/archive/IntCal/IntCal20/index.json")
)
}
\arguments{
\item{fname}{The name of the json file. By default this downloads the json file from the intchron.org site}
}
\description{
Download the json file that contains the IntCal20 radiocarbon calibration curves and the contributing data series.
}
\details{
The intcal curves consist of the IntCal20, SHCal20 and Marine20 calibration curves. The details of these curves can be loaded, as well as the underlying data such as tree-ring records.
}
\examples{
 intcal <- intcal.read.data() # from intchron.org
}
