% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcal_json.R
\name{intcal.write.data}
\alias{intcal.write.data}
\title{Write intcal data to a file.}
\usage{
intcal.write.data(data, fname)
}
\arguments{
\item{data}{intcal variable as obtained from intcal.read.data()}

\item{fname}{Name of the file to be written}
}
\description{
Write the intcal.json file that comes with the rintcal packages to somewhere local. This can be useful if you want to avoid repeatedly downloading the json file from intchron.org.
}
\examples{
 intcal <- intcal.read.data()
 myintcal <- tempfile()
 intcal.write.data(intcal, myintcal)
}
