% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rintcal.R
\name{list.ccurves}
\alias{list.ccurves}
\title{List the calibration curves}
\usage{
list.ccurves()
}
\value{
A list of the available calibration curves
}
\description{
List the file names of the calibration curves available within the rintcal package.
For details such as references of the curves, see the documentation of the function `ccurve()`.
}
