% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntCal.R
\name{new.ccdir}
\alias{new.ccdir}
\title{Make directory and fill with calibration curves}
\usage{
new.ccdir(ccdir = "./ccurves")
}
\arguments{
\item{ccdir}{Name and location of the new directory. Defaults to `ccurves', a folder within the current working directory, \code{ccdir="./ccurves"}.}
}
\description{
Make an alternative `curves' directory and fill it with the calibration curves.
}
\details{
Copies all calibration curves within the `IntCal' package to the new directory.
}
\examples{
new.ccdir(tempdir())
}
