% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rintcal-package.R
\docType{package}
\name{rintcal-package}
\alias{rintcal}
\alias{rintcal-package}
\title{rintcal: Radiocarbon Calibration Curves}
\description{
The IntCal20 radiocarbon calibration curves (Reimer et al. 2020 \doi{10.1017/RDC.2020.68}) are provided as a data package, together with previous IntCal curves (IntCal13, IntCal09, IntCal04, IntCal98), other curves (e.g., NOTCal04 [van der Plicht et al. 2004], Arnold & Libby 1951, Stuiver & Suess 1966, Pearson & Stuiver 1986) and postbomb curves. Also provided are functions to copy the curves into memory, and to read, query and plot the data underlying the IntCal20 curves.
}
\author{
\strong{Maintainer}: Maarten Blaauw \email{maarten.blaauw@qub.ac.uk} (\href{https://orcid.org/0000-0002-5680-1515}{ORCID})

Authors:
\itemize{
  \item Maarten Blaauw \email{maarten.blaauw@qub.ac.uk} (\href{https://orcid.org/0000-0002-5680-1515}{ORCID})
}

Other contributors:
\itemize{
  \item Christopher Bronk Ramsey \email{christopher.ramsey@arch.ox.ac.uk} (\href{https://orcid.org/0000-0002-8641-9309}{ORCID}) [contributor]
}

}
\keyword{internal}
