% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rintcal.R
\name{glue.ccurves}
\alias{glue.ccurves}
\title{Glue prebomb and postbomb curves}
\usage{
glue.ccurves(prebomb = "IntCal20", postbomb = "NH1")
}
\arguments{
\item{prebomb}{The prebomb curve. Defaults to "IntCal20"}

\item{postbomb}{The postbomb curve. Defaults to "NH1" (Hua et al. 2013)}
}
\value{
The custom-made curve (invisibly)
}
\description{
Produce a custom curve by merging two calibration curves, e.g. a prebomb and a postbomb one for dates which straddle both curves.
}
\examples{
my.cc <- glue.ccurves()
}
