% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{hpd}
\alias{hpd}
\title{Calculate highest posterior density}
\usage{
hpd(caldist, prob = 0.95, hpdsteps = 1, yrsteps = FALSE, rule = 1)
}
\arguments{
\item{caldist}{The calibrated distribution, as returned from caldist()}

\item{prob}{Probability range which should be calculated. Default \code{prob=0.95}.}

\item{hpdsteps}{It works.}

\item{yrsteps}{This too. Could be used e.g. for postbomb dates?}

\item{rule}{Which extrapolation rule to use. Defaults to \code{rule=1} which returns NAs.}
}
\description{
Calculate highest posterior density ranges of calibrated distribution
}
\examples{
hpd(caldist(130,20))
}
