% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntCal.R
\name{intcal.data}
\alias{intcal.data}
\title{plot the IntCal20 data}
\usage{
intcal.data(
  calmin,
  calmax,
  cc1 = "IntCal20",
  cc2 = NA,
  BCAD = FALSE,
  cal.lab = NA,
  cal.rev = FALSE,
  c14.lab = NA,
  c14.lim = NA,
  c14.rev = FALSE,
  ka = FALSE,
  cc1.col = rgb(0, 0, 1, 0.5),
  cc1.fill = rgb(0, 0, 1, 0.2),
  cc2.col = rgb(0, 0, 0.5, 0.5),
  cc2.fill = rgb(0, 0, 0.5, 0.2),
  data.cols = 1:8,
  data.pch = c(1, 2, 5, 6, 15:19),
  pch.cex = 0.5,
  legend.loc = "topleft",
  legend.ncol = 2,
  legend.cex = 0.7,
  cc.legend = "bottomright",
  bty = "l",
  ...
)
}
\arguments{
\item{calmin}{Minimum calendar year for the plot}

\item{calmax}{Maximum calendar year for the plot}

\item{cc1}{Name of the calibration curve. Can be "IntCal20", "Marine20", "SHCal20", or for the previous curves "IntCal13", "Marine13" or "SHCal13".}

\item{BCAD}{The calendar scale of graphs and age output-files is in cal BP (calendar or calibrated years before the present, where the present is AD 1950) by default, but can be changed to BC/AD using \code{BCAD=TRUE}.}

\item{cal.lab}{The labels for the calendar axis (default \code{age.lab="cal BP"} or \code{"BC/AD"} if \code{BCAD=TRUE}), or to \code{age.lab="kcal BP"} etc. if ka=TRUE.}

\item{cal.rev}{Reverse the calendar axis.}

\item{c14.lab}{Label for the C-14 axis. Defaults to 14C BP (or 14C kBP if ka=TRUE).}

\item{c14.lim}{Axis limits for the C-14 axis. Calculated automatically by default.}

\item{c14.rev}{Reverse the C-14 axis.}

\item{ka}{Use kcal BP (and C14 kBP).}

\item{cc1.col}{Colour of the calibration curve (outline).}

\item{cc1.fill}{Colour of the calibration curve (fill).}

\item{cc2.col}{Colour of the calibration curve (outline), if activated (default cc2=NA).}

\item{cc2.fill}{Colour of the calibration curve (fill), if activated (default cc2=NA).}

\item{data.cols}{colours of the data points. Defaults to R's colours 1 to 8 (black, red, green, darkblue, lightblue, purple, orange, and grey)}

\item{data.pch}{Symbols of the data points. Defaults to R's symbols 1, 2, 5, 6, and 15 to 19 (open circle, open upward triangle, open diamond, open downward triangle, closed square, closed circle, closed upward triangle, closed diamond)}

\item{pch.cex}{Size of the data symbols. Defaults to 0.5.}

\item{legend.loc}{Location of the data legend. Defaults to topleft. Set to NA for no plotting.}

\item{legend.ncol}{Number of columns of the data legend.}

\item{legend.cex}{Size of the legend. Defaults to 0.7.}

\item{cc.legend}{Location of the legend for the calibration curve(s).}

\item{bty}{Box type around the plot. Defaults to "l"-shaped.}

\item{...}{Any additional optional plotting parameters.}

\item{cc2.}{Optional second calibration curve to plot. Can be "IntCal20", "Marine20", "SHCal20", or for the previous curves "IntCal13", "Marine13" or "SHCal13". Defaults to nothing, NA.}
}
\description{
plot the C14 ages underpinning the IntCal20/Marine20/SHCal20 calibration curves
}
\details{
These datasets were downloaded from Intcal.org. All data have both uncertainties in C14 age and on the calendar scale. For trees this is the sample thickness (e.g., 10 years or 1 year).
}
\examples{
  intcal.data(100, 200)
  intcal.data(40e3, 55e3, ka=TRUE)
}
