% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntCal.R
\docType{package}
\name{IntCal}
\alias{IntCal}
\title{IntCal}
\description{
The international IntCal research group publishes ratified radiocarbon calibration curves such as IntCal20, Marine20 and SHCal20 (Reimer et al. 2020).
This data package provides the files of these curves, for use by other R package (reducing the need for replication and the size of other packages that use IntCal).
It also comes with a limited number of relevant functions, to read in calibration curves, translate pMC ages to 14C ages (et vice versa), etc.
}
\author{
Maarten Blaauw <maarten.blaauw@qub.ac.uk>
}
