% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rintcal.R
\name{new.ccdir}
\alias{new.ccdir}
\title{Make directory and fill with calibration curves}
\usage{
new.ccdir(ccdir)
}
\arguments{
\item{ccdir}{Name and location of the new directory. For example, this could be a folder called 'ccurves', living within the current working directory, \code{ccdir="./ccurves"}.}
}
\value{
A message informing the user the name of the folder into which the calibration curves have been copied.
}
\description{
Make an alternative `curves' directory and fill it with the calibration curves.
}
\details{
Copies all calibration curvccdires within the `rintcal' package to the new directory.
}
\examples{
new.ccdir(tempdir())
}
